--- src/main/kotlin/edu/example/kotlindevelop/domain/product/product/repository/ProductRepository.kt
@@ -1,10 +1,0 @@
-package edu.example.kotlindevelop.domain.product.product.repository
-
-import edu.example.kotlindevelop.domain.product.product.entity.Product
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface ProductRepository : JpaRepository<Product?, Long?> {
-
-}
