+++ src/main/kotlin/edu/example/kotlindevelop/domain/product/repository/LossRateRepository.kt
@@ -1,0 +1,8 @@
+package edu.example.kotlindevelop.domain.product.repository
+
+import edu.example.kotlindevelop.domain.product.entity.LossRate
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface LossRateRepository : JpaRepository<LossRate, Long> {
+
+}
