+++ src/main/kotlin/edu/example/kotlindevelop/domain/product/repository/ProductRepository.kt
@@ -2,6 +2,9 @@ package edu.example.kotlindevelop.domain.product.repository
 
 import edu.example.kotlindevelop.domain.product.entity.Product
 import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
 
-interface ProductRepository : JpaRepository<Product, Long>{
+@Repository
+interface ProductRepository : JpaRepository<Product?, Long?> {
+
 }
